Include panic-halt crate
println is a macro found in the std crate. We don't need it at the
moment, so we can remove it and try to build again.

Build 3

$ cargo build --target thumbv6m-none-eabi
error: `#[panic_handler]` function required, but not found
This error, is because rustc required a panic handler to be implemented.

panic_impl

We could try and implement the panic macro ourselves, but it's easier
and more portable to use a crate that does it for us.

	modified:   Cargo.toml
	modified:   src/main.rs