error: requires `start` lang_item
no_main

In the normal command line rust binaries you would be used to making,
executing the binary usually has the operating system start by executing
the C runtime library (crt0). This in turn invokes the Rust runtime, as
marked by the start language item, which in turn invokes the main
function.

Having enabled no_std, as we are targeting on a microcontroller, neither
the crt0 nor the rust runtime are available, so even implementing start
would not help us. We need to replace the operating system entry point.

You could for example name a function after the default entry point,
which for linux is _start, and start that way. Note, you would also need
to disable name mangling:

	modified:   Cargo.lock
	modified:   src/main.rs